+++ src/components/english/HomeObject.tsx
@@ -1,0 +1,3 @@
+import '../../i18n/english';
+import HomeObject from '../HomeObject';
+export default HomeObject;